requirements.txt
python-dotenv>=1.0.1
requests>=2.31.0
pandas>=2.2.2
numpy>=1.26.4
yfinance>=0.2.43
nltk>=3.8.1
tabulate>=0.9.0
tqdm>=4.66.4
langchain>=0.2.12
langchain-community>=0.2.11
langchain-core>=0.2.35
langchain-openai>=0.1.21
openai>=1.35.7
rich>=13.7.1
streamlit>=1.36.0
pytest>=8.2.0
